Internal: Unify editor loader into single module [ED-24788]
Ref: ED-24788

Collapse factory, interface, abstract base, and V2 loader into one
Editor_Loader class. Move assets out of loader/v2/ and drop -v2 from
internal JS handles and bundle names. Preserve elementor/editor/v2/*
hooks and #elementor-editor-wrapper-v2 for Pro compatibility.
