+++ .grunt-config/webpack.js
@@ -125,8 +125,8 @@ const entry = {
 	'interactions': path.resolve( __dirname, '../modules/interactions/assets/js/interactions.js' ),
 	'editor-interactions': path.resolve( __dirname, '../modules/interactions/assets/js/editor-interactions.js' ),
 	'kit-elements-defaults-editor': path.resolve( __dirname, '../modules/kit-elements-defaults/assets/js/editor/index.js' ),
-	'editor-loader-v2': path.resolve( __dirname, '../core/editor/loader/v2/js/editor-loader-v2.js' ),
-	'editor-environment-v2': path.resolve( __dirname, '../core/editor/loader/v2/js/editor-environment-v2.js' ),
+	'editor-loader': path.resolve( __dirname, '../core/editor/loader/js/editor-loader.js' ),
+	'editor-environment': path.resolve( __dirname, '../core/editor/loader/js/editor-environment.js' ),
 	'responsive-bar': path.resolve( __dirname, '../assets/dev/js/editor/regions/responsive-bar/index.js' ),
 	'ai': path.resolve( __dirname, '../modules/ai/assets/js/editor/index.js' ),
 	'admin-notifications': path.resolve( __dirname, '../modules/notifications/assets/js/admin.js' ),
