+++ core/editor/loader/js/editor-environment.js
@@ -2,4 +2,4 @@ if ( ! window.elementorV2?.env ) {
 	throw new Error( 'The "@elementor/env" package was not loaded.' );
 }
 
-window.elementorV2.env.initEnv( window.elementorEditorV2Env );
+window.elementorV2.env.initEnv( window.elementorEditorEnv );
