+++ core/editor/loader/js/editor-loader.js
