+++ core/editor/loader/scss/editor-v2-app-bar-overrides.scss
